# adapter/adapter.rest/adapter.rest.v2/src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "https://github.com/learning-layers/SocialSemanticServer/",
    "version" : "v2",
    "title" : "Social Semantic Server REST API",
    "termsOfService" : "https://github.com/learning-layers/",
    "contact" : {
      "email" : "dtheiler#tugraz$at"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/sss.adapter.rest.v2/rest",
...
  "paths" : {
...
    "/friends/friends/{friend}" : {
      "post" : {
        "tags" : [ "friends" ],
        "summary" : "add a friend",
        "description" : "",
        "operationId" : "friendAddPost",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "friend",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSFriendAddRet"
            }
          }
        }
      }
    },
    "/images/images/profile/picture/entity/{entity}/file/{file}" : {
      "put" : {
        "tags" : [ "images" ],
        "summary" : "set an entity's profile picture",
        "description" : "",
        "operationId" : "imageProfilePictureSet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "entity",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "file",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSImageProfilePictureSetRet"
            }
          }
        }
      }
    },
...
    "/likes/likes/entities/{entity}/value/{value}" : {
      "put" : {
        "tags" : [ "likes" ],
        "summary" : "like / dislike / neutral an entity",
        "description" : "",
        "operationId" : "likeUpdate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "entity",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "value",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLikeUserSetRet"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SSFriendAddRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "friend" : {
          "type" : "string"
        }
      }
    },
...
    "SSImageProfilePictureSetRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "user" : {
          "type" : "string"
        }
      }
    },
...
    "SSLikeUserSetRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "entity" : {
          "type" : "string"
        }
      }
    },
...
  }
}